+++ src/renderer/components/dialogs/Settings-Account.tsx
@@ -2,61 +2,69 @@ import { useEffect, useState, useContext } from 'react'
 import { DeltaBackend } from '../../delta-remote'
 import { Card, Elevation } from '@blueprintjs/core'
 import React from 'react'
-import LoginForm, { ConfigureProgressDialog } from '../LoginForm'
+import LoginForm, {
+  ConfigureProgressDialog,
+  defaultCredentials,
+} from '../LoginForm'
 
 import { DeltaDialogBody, DeltaDialogOkCancelFooter } from './DeltaDialog'
-import { ScreenContext, useTranslationFunction } from '../../contexts'
+import { ScreenContext } from '../../contexts'
+import { Credentials } from '../../../shared/shared-types'
 
 export default function SettingsAccount({
   setShow,
+  onClose,
 }: {
   show: string
   setShow: (show: string) => void
-  onClose: any
+  onClose: () => void
 }) {
-  const [, setInitialAccountSettings] = useState<{
-    [key: string]: string
-  }>(null)
+  const [, setInitialAccountSettings] = useState<Credentials>(
+    defaultCredentials()
+  )
 
-  const [accountSettings, _setAccountSettings] = useState<{
-    [key: string]: string
-  }>(null)
+  const [accountSettings, _setAccountSettings] = useState<Credentials>(
+    defaultCredentials()
+  )
 
   const [disableUpdate, setDisableUpdate] = useState(true)
 
-  const setAccountSettings = (
-    value: React.SetStateAction<{
-      [key: string]: string
-    }>
-  ) => {
+  const setAccountSettings = (value: Credentials) => {
     disableUpdate === true && setDisableUpdate(false)
     _setAccountSettings(value)
   }
 
-  const { openDialog, userFeedback } = useContext(ScreenContext)
-  const tx = useTranslationFunction()
+  const { openDialog } = useContext(ScreenContext)
 
   const loadSettings = async () => {
-    const accountSettings = await DeltaBackend.call('settings.getConfigFor', [
-      'addr',
-      'mail_pw',
-      'inbox_watch',
-      'sentbox_watch',
-      'mvbox_watch',
-      'mvbox_move',
-      'e2ee_enabled',
-      'mail_server',
-      'mail_user',
-      'mail_port',
-      'mail_security',
-      'imap_certificate_checks',
-      'send_user',
-      'send_pw',
-      'send_server',
-      'send_port',
-      'send_security',
-      'smtp_certificate_checks',
-    ])
+    const accountSettings: Credentials = ((await DeltaBackend.call(
+      'settings.getConfigFor',
+      [
+        'addr',
+        'mail_pw',
+        'inbox_watch',
+        'sentbox_watch',
+        'mvbox_watch',
+        'mvbox_move',
+        'e2ee_enabled',
+        'mail_server',
+        'mail_user',
+        'mail_port',
+        'mail_security',
+        'imap_certificate_checks',
+        'send_user',
+        'send_pw',
+        'send_server',
+        'send_port',
+        'send_security',
+        'smtp_certificate_checks',
+        'socks5_enabled',
+        'socks5_host',
+        'socks5_port',
+        'socks5_user',
+        'socks5_password',
+      ]
+    )) as unknown) as Credentials
     setInitialAccountSettings(accountSettings)
     _setAccountSettings(accountSettings)
   }
@@ -67,7 +75,8 @@ export default function SettingsAccount({
 
   const onUpdate = () => {
     if (disableUpdate) return
-    const onSuccess = () => userFeedback({ type: 'success', text: tx('ok') })
+    const onSuccess = () => onClose()
+
     openDialog(ConfigureProgressDialog, {
       credentials: accountSettings,
       onSuccess,