+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [Unreleased][unreleased]
 
+### Added
+
+- add switch to toggle simultaneous account syncing off
 ## [1.21.0] - 2021-09-08
 
 ### Added