+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## [Unreleased][unreleased]
 
+### Added
+- Implement settings for socks5 proxy
+
+
+### Fixed
+- fix yggmail emailaddress text overflow in qrcode dialog & settings
 
 ## [1.21.0] - 2021-09-08
 
@@ -13,7 +19,6 @@
 - "rocket chat"-like experimental theme
 - Implement new contact request ui/ux
 - Implement new connectivity view
-- Implement settings for socks5 proxy
 
 ### Changed
 - Update translations (06.09.2021)