+++ CHANGELOG.md
@@ -3,12 +3,16 @@
 ## [Unreleased][unreleased]
 
 ### Added
-
 - add switch to toggle simultaneous account syncing off
+- Implement settings for socks5 proxy
 
 ### Fixed
 
 - fix flashing up account list on startup
+
+
+### Fixed
+- fix yggmail emailaddress text overflow in qrcode dialog & settings
 
 ## [1.21.0] - 2021-09-08