+++ CHANGELOG.md
@@ -2,9 +2,14 @@
 
 ## [Unreleased][unreleased]
 
-### Fixed
+### Added
+- add switch to toggle simultaneous account syncing off
+- Implement settings for socks5 proxy
 
+### Fixed
+- fix flashing up account list on startup
 - fix update/load core translation strings
+- fix yggmail emailaddress text overflow in qrcode dialog & settings
 
 ## [1.21.0] - 2021-09-08